Add pretty_json and node8
Use `command -v` to check if node already exists on machine.

Check if TTY in node8.

Playing with JS-style that is very DRY, functional, and has narrow variable scopes.
The result seems to be a Pyramid of Doom, where radability becomes a
ping-pong between upper and lower lines, but otherwise is pleasant :D