Enable mouse click
Click and release events are available, in addition to button identification and mofifier keys pressed:
https://invisible-island.net/xterm/ctlseqs/ctlseqs.html#h2-Mouse-Tracking
https://www.xfree86.org/current/ctlseqs.html#Mouse%20Tracking

The problem is that terminal text selection doesn't work.

I.e. great mouse support should be implemented,
and a way to easily disable/enable mouse capturing?

Status bar with [x]Mouse which can be de-selected with click or key?
With this, great mouse support is less important.

What is great mouse support?
Selecting range of commits and provide useful actions? Also a way to copy the hashes?
Selecting code lines?