Page by (committer) time
Instead of commit hash, so that all branches are covered.

For this to work, graph must be ordered by committer date (`--date-order`).
Author date is of course not trustworthy because of rebase, cherry-pick, etc.
Default ordering is apparently:
> try to show a commit on a single-parent chain together
https://github.com/git/git/blob/f443b226ca681d87a3a31e245a70e6bc2769123c/commit.h#L221-L235

To be less confusing, local format of committer date is used, and committer date is displayed in commit (`--pretty=fuller`)

Because commits can have identical timestamps, I increase the commit index (>0) when appropriate.